Update .travis.yml and generate-apk.sh

+++ generate-apk.sh
@@ -1,48 +1,53 @@
 #!/bin/bash
 
-if [ "$TRAVIS_PULL_REQUEST" != "false" -o "$TRAVIS_REPO_SLUG" != "m-murad/keep" ]; then
+# Travis build triggered on a forked repository
+if [ "$TRAVIS_REPO_SLUG" != "m-murad/keep" ]; then
+    echo "Not the original repo. Skip apk upload."
+    exit 0
+fi
+
+# Travis build triggered by a PR
+if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
     echo "Just a PR. Skip apk upload."
     exit 0
 fi
 
-# Set git username
-git config --global user.name "m-murad"
+# Travis build triggered by commit to a branch other than master
+if [ "$TRAVIS_BRANCH" != "master" ]; then
+    echo "Not pushed to main branch. Skip apk upload."
+    exit 0
+fi
 
-# Set git email
-git config --global user.email "murad.kuka@gmail.com"
-
-cd $HOME
-
-# Clone the "apk" branch of the repository
-git clone --quiet --branch=apk https://m-murad:$GITHUB_API_KEY@github.com/m-murad/keep  apk > /dev/null
-
-# cd to the cloned repository
-cd apk
-
-# Create an orphan branch with no initial commit
-git checkout --orphan temp
-
-# Delete the previous apk(s) 
-rm -rf $HOME/apk/*
-
-# Unstage all the files in your working tree.
-git rm --cached $(git ls-files)
-
-# Copy newly generated apk(s) to current directory
-cp $HOME/build/m-murad/keep/app/build/outputs/apk/debug/app-debug.apk $HOME/apk/
-cp $HOME/build/m-murad/keep/app/build/outputs/apk/release/app-release-unsigned.apk $HOME/apk/
-
-# Add and commit the files
-git add -f app-debug.apk
-git add -f app-release-unsigned.apk
-git add .travis.yml
-git commit --message "Apk(s) update for build:$TRAVIS_BUILD_NUMBER"
-
-# Delete the branch name "apk"
-git branch -D apk
-
-# Rename the current branch from "temp" to "apk"
-git branch -m apk
-
-# Force push to the apk branch
-git push origin apk --force --quiet> /dev/null
+# Travis build triggered by commit to master branch
+if [ "$TRAVIS_BRANCH" == "master" ]; then
+    echo "Pushed to master branch. Uploading apk(s) to apk branch"
+    # Set git username
+    git config --global user.name "m-murad"
+    # Set git email
+    git config --global user.email "murad.kuka@gmail.com"
+    cd $HOME
+    # Clone the "apk" branch of the repository
+    git clone --quiet --branch=apk https://m-murad:$GITHUB_API_KEY@github.com/m-murad/keep  apk > /dev/null
+    # cd to the cloned repository
+    cd apk
+    # Create an orphan branch with no initial commit
+    git checkout --orphan temp
+    # Delete the previous apk(s)
+    rm -rf $HOME/apk/*
+    # Unstage all the files in your working tree
+    git rm --cached $(git ls-files)
+    # Copy newly generated apk(s) to current directory
+    cp $HOME/build/m-murad/keep/app/build/outputs/apk/debug/app-debug.apk $HOME/apk/
+    cp $HOME/build/m-murad/keep/app/build/outputs/apk/release/app-release-unsigned.apk $HOME/apk/
+    # Add and commit the files
+    git add -f app-debug.apk
+    git add -f app-release-unsigned.apk
+    git add .travis.yml
+    git commit --message "Apk(s) update for build:$TRAVIS_BUILD_NUMBER"
+    # Delete the branch name "apk"
+    git branch -D apk
+    # Rename the current branch from "temp" to "apk"
+    git branch -m apk
+    # Force push to the apk branch
+    git push origin apk --force --quiet> /dev/null
+fi
